start removing old gpaw interface
