+++ .gitignore
@@ -1,6 +1,8 @@
 .idea/
 .vscode/
 build/
+ciderpress/lib/deps/
+ciderpress/lib/pwutil/config.h
 dist/
 docs/_build
 docs/_static
